+++ src/main/java/com/example/Pastach/exception/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.Pastach.exception;
 
 public class PostNotFoundException extends RuntimeException {
-    public PostNotFoundException(int id) {
+    public PostNotFoundException(Long id) {
         super("Post with id " + id + " not found");
     }
 }
